Don't lower file handle limit if sufficient
The latest ct-ng wants 2048 handles. It seems silly to lower it if it starts
out higher so this makes setting it conditional on it being too low to
begin with.

This code was almost copy/pasted from the latest ct-ng and can be
removed when we've stopped using old versions.