Enable Fortran
Fortran support increases archive size by about 15% (~10 MB) and adds a
shared library.

The Fortran runtime is not added to firmware images unless enabled by
Buildroot. The official Nerves systems do not enable it. However, if you
want it, enabling it in the system is way easier than enabling it in
both a custom toolchain and the system.