Workaround for argument overflow due to long paths
This moves the ctng build directory to /tmp to significantly trim path
lengths.

The work directory previously had been things like:

$HOME/toolchains/nerves_toolchain_arm_unknown_linux_gnueabihf/.nerves/artifacts/nerves_toolchain_arm_unknown_linux_gnueabihf-linux_x86_64-1.3.0/build/.build

It's now /tmp/ct-work.

The other option of fixing the code that caused the buffer overflow
looked difficult since it was in a gcc Makefile.

The downside of this fix is that it forces only one toolchain build at a
time. This seems like a reasonable tradeoff given how parallelized the
toolchain builds already are.