Linux: Bump headers to 4.14
This forces systems to use Linux 4.14 or later. The official ones
already do, though. The benefit is that programs will now be able to use
syscalls introduced after 4.4. This majorly improves eBPF availability,
but also brings in a few smaller ones that some programs might want -
like zero-copy on sockets.