nerves_toolchain_armv7_nerves_linux_gnueabihf: new toolchain
This is a replacement for `nerves_toolchain_arm_unknown_linux_gnueabihf`
based on the nearly equivalent Bootlin configuration for ARMv7-A cores.
So that it works on both Cortex A8 and Cortex A9 processors, Bootlin
removed neon support and limited floating point to vfpv3-d16 (instead of
the 32-register vfpv3).

[no-ci]