Bump Linux headers back to 4.14
The Linux headers were backed off to 4.4 in
75287aa164529e23513440cc09d5d2f11b783813. This was done to increase
compatibility of the Nerves toolchains with vendor supplied kernels.

It turns out that use of cdev and eBPF were impacted by this change.
The minimum kernel header version for both of these is 4.14.

The only Nerves platform that's affect are the NVidia Jetson Nanos. This
is not an officially supported platform.