Switch default ARMv7 CPU to a generic one
I believe that I misunderstood the context of why the Bootlin
configuration for their ARMv7 compiler specified a Cortex-A9. I think
that it uses a generic ARMv7 configuration behind the scenes. This may
explain why we started seeing the Linux kernel handle at least NEON
instruction on the AM3358 (Cortex-A8) rather than it being handled
directly by the processor when we changed to this compiler.