Switch RISC-V from Glibc to Musl Libc
Glibc has two issues that make it unusable for now in RISC-V:

1. The `start_thread` function in segfaults. This prevents any programs
   using pthread from creating a thread.
2. It doesn't seem to be possible to turn off multilib. This means
   libraries are stored both under `/lib` and `/lib64`. This doesn't
   seem to be handled by Buildroot's external toolchain library finder.

To work around this, use Musl libc which doesn't have either issue.

This deprecated the glibc toolchain since it's already been published to
hex.pm. Presumably these issues will be fixed at some point in the
future and we may undeprecate this then.