Don't link with zstd
libzstd is getting pulled in even though it's unused. Normally, there's
a way to turn this off if you're using LTO. We're not using LTO since
it's disabled due to building static toolchains.

See https://github.com/crosstool-ng/crosstool-ng/issues/1579 for some
information.

It appears that libzstd is the only library that causes the toolchains
to depend on Homebrew. Hopefully this:

Fixes #114