Let gcc determine whether to use SJLJ
This option was inconsistently used. Removing it entirely makes all
toolchains the same.

The note for this option says that it's safe to let it detect whether
setjmp/longjmp is needed for unwinding stacks for exceptions. The only
caveat was a note about a big endian platform which doesn't affect us.