Remove static toolchain option since inconsistent
It turns out that MacOS ignored it since configure detected that the
linker didn't support it. Therefore, static compilation was turned off
AND other options got automatically turned on. One of those was LTO.
There might be others.

Since the flag to prevent libzstd from being linked was in response to
configure automatically doing this even with LTO was off, turn on the
crosstool-ng flag that prevents it.