Bump version to follow gcc
This does a pretty major version bump to make it much easier to figure
out what gcc version is being used. This question comes up a lot.

The new version format is expected to be gcc_major.gcc_minor.patch where
the patch could be due to any changes - crosstool-ng, gcc, Nerves, etc.
Semver will likely be violated a little with this versioning change, but
backwards incompatible changes will need to be dealt with differently.