Bump kernel headers from 4.19 to 5.4
Linux 5.10 is now the oldest in use by official Nerves Systems. It's
possible that Linux 5.4 is on some systems in the wild. Earlier versions
of Linux probably will be stuck using GCC 12 and earlier so this
probably isn't a big issue. On the plus side, it enables access to newer
kernel APIs.